requirements.txt
wheel==0.44.0
python-dotenv==1.0.1

steam==1.4.4
protobuf==3.20.3
gevent==24.10.3
gevent-eventemitter==2.1
google-api-python-client==2.149.0

ms_cv==0.1.1

discord-webhook==1.3.1

legendary-gl==0.20.34

tabulate==0.9.0

python-dotenv
discord-webhook